backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
psycopg2-binary
openpyxl
requests
gunicorn
Flask-CORS
pytest
